requirements.txt
Pillow==10.2.0
azure-storage-blob==12.19.0 